requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
email-validator==2.2.0
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
pytest==8.3.2
pydantic-settings==2.5.2
httpx==0.27.2
loguru==0.7.2
sqlalchemy-schemadisplay==1.3
graphviz==0.20.3
pydot==2.0.0
jinja2==3.1.4
openpyxl==3.1.5
reportlab==4.2.2
redis==5.0.8
rq==1.16.2
celery==5.4.0
boto3==1.35.46
fakeredis==2.23.3
cryptography==43.0.1
prometheus-client==0.20.0


# Google Drive backend
google-auth==2.29.0
google-api-python-client==2.126.0
# Outils qualité / tests avancés
# ruff==0.6.9
ruff==0.12.8
mypy==1.11.2
# pytest-cov==5.0.0
pytest-cov==6.2.1
types-redis==4.6.0.20241004
types-requests==2.32.0.20241016
